split component batch adapter headers
