+++ rerun_cpp/src/rerun/archetypes/arrows3d.cpp
@@ -3,6 +3,8 @@
 
 #include "arrows3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Arrows3D::INDICATOR_COMPONENT_NAME[] = "rerun.components.Arrows3DIndicator";
