+++ rerun_cpp/src/rerun/archetypes/bar_chart.cpp
@@ -3,6 +3,8 @@
 
 #include "bar_chart.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char BarChart::INDICATOR_COMPONENT_NAME[] = "rerun.components.BarChartIndicator";
