+++ rerun_cpp/src/rerun/archetypes/boxes3d.cpp
@@ -3,6 +3,8 @@
 
 #include "boxes3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Boxes3D::INDICATOR_COMPONENT_NAME[] = "rerun.components.Boxes3DIndicator";
