+++ rerun_cpp/src/rerun/archetypes/boxes3d_ext.cpp
@@ -1,5 +1,7 @@
 #include "boxes3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp" // TODO(#3991): Should not leak into public header.
+
 // #define EDIT_EXTENSION
 
 namespace rerun {
