+++ rerun_cpp/src/rerun/archetypes/clear.cpp
@@ -3,6 +3,8 @@
 
 #include "clear.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Clear::INDICATOR_COMPONENT_NAME[] = "rerun.components.ClearIndicator";
