+++ rerun_cpp/src/rerun/archetypes/depth_image.cpp
@@ -3,6 +3,8 @@
 
 #include "depth_image.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char DepthImage::INDICATOR_COMPONENT_NAME[] = "rerun.components.DepthImageIndicator";
