+++ rerun_cpp/src/rerun/archetypes/line_strips3d.cpp
@@ -3,6 +3,8 @@
 
 #include "line_strips3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char LineStrips3D::INDICATOR_COMPONENT_NAME[] =
