+++ rerun_cpp/src/rerun/archetypes/points3d.cpp
@@ -3,6 +3,8 @@
 
 #include "points3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Points3D::INDICATOR_COMPONENT_NAME[] = "rerun.components.Points3DIndicator";
