+++ rerun_cpp/src/rerun/archetypes/segmentation_image.cpp
@@ -3,6 +3,8 @@
 
 #include "segmentation_image.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char SegmentationImage::INDICATOR_COMPONENT_NAME[] =
