+++ rerun_cpp/src/rerun/archetypes/tensor.cpp
@@ -3,6 +3,8 @@
 
 #include "tensor.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Tensor::INDICATOR_COMPONENT_NAME[] = "rerun.components.TensorIndicator";
