+++ rerun_cpp/src/rerun/archetypes/text_document.cpp
@@ -3,6 +3,8 @@
 
 #include "text_document.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char TextDocument::INDICATOR_COMPONENT_NAME[] =
