+++ rerun_cpp/src/rerun/archetypes/time_series_scalar.cpp
@@ -3,6 +3,8 @@
 
 #include "time_series_scalar.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char TimeSeriesScalar::INDICATOR_COMPONENT_NAME[] =
