+++ rerun_cpp/src/rerun/archetypes/transform3d.cpp
@@ -3,6 +3,8 @@
 
 #include "transform3d.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char Transform3D::INDICATOR_COMPONENT_NAME[] =
