+++ rerun_cpp/src/rerun/archetypes/view_coordinates.cpp
@@ -3,6 +3,8 @@
 
 #include "view_coordinates.hpp"
 
+#include "../component_batch_adapter_builtins.hpp"
+
 namespace rerun {
     namespace archetypes {
         const char ViewCoordinates::INDICATOR_COMPONENT_NAME[] =
