+++ rerun_cpp/src/rerun/serialized_component_batch.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "data_cell.hpp"
 
 namespace rerun {
